Check description
Reworked environment, should be more performant.
Added housing and distortion effect to Christian's portal.
Added door that can be opened to initiate Marcus's fade in and a cube for the kitchen.

To do -

change gizmo in fade in script to not be instanced so we can have separate gizmos for kitchen and living room.

Portals are not behaving correctly. eg not spawning in correct location in level 2.

Integrate TV puzzle and inventory system and start on kitchen puzzle